MAINT: Remove debye folder, move debye_gruneisen.py up one level, and rename to debye.py
